+++ src/importer/meta.rs
@@ -8,11 +8,8 @@ use std::{
     path::Path,
 };
 
-use graphannis::{
-    model::AnnotationComponentType,
-    update::{GraphUpdate, UpdateEvent},
-};
-use graphannis_core::{graph::ANNIS_NS, util::split_qname};
+use graphannis::update::{GraphUpdate, UpdateEvent};
+use graphannis_core::util::split_qname;
 use serde_derive::Deserialize;
 
 use crate::{progress::ProgressReporter, util::get_all_files, Module, StepID};
@@ -64,56 +61,28 @@ impl Importer for AnnotateCorpus {
         let mut update = GraphUpdate::default();
         let all_files = get_all_files(input_path, vec!["meta"])?;
         let progress = ProgressReporter::new(tx, step_id, all_files.len())?;
-        for file_path in all_files {
-            let mut corpus_nodes = Vec::new();
-            for ancestor in file_path.ancestors() {
-                if ancestor == input_path {
-                    break;
-                }
-                corpus_nodes.push(ancestor);
-            }
-            corpus_nodes.reverse();
-            let last_item = corpus_nodes.remove(corpus_nodes.len() - 1);
-            let clean_name = last_item
-                .to_path_buf()
-                .parent()
-                .unwrap()
-                .join(last_item.file_stem().unwrap());
-            corpus_nodes.push(clean_name.as_path());
-            let start_index: usize = input_path.to_str().unwrap().len() + 1;
-            let mut previous: Option<String> = None;
-            for node_path in corpus_nodes {
-                let node_name = (node_path.to_str().unwrap()[start_index..]).to_string();
-                update.add_event(UpdateEvent::AddNode {
+        let start_index = input_path.to_string_lossy().len() + 1;
+        for file_path in all_files.into_iter().filter(|p| p.is_file()) {
+            let parent = &file_path.parent().unwrap();
+            let file_stem = file_path.file_stem().unwrap();
+            let full_path = &parent.join(file_stem);
+            let node_name = &full_path.to_string_lossy()[start_index..];
+            update.add_event(UpdateEvent::AddNode {
+                node_name: node_name.to_string(),
+                node_type: "corpus".to_string(),
+            })?; // this is required, corpus annotations might be first updates to be processed
+            let annotations = read_annotations(&file_path, &progress)?;
+            for (k, v) in annotations {
+                let (anno_ns, anno_name) = match split_qname(k.as_str()) {
+                    (None, name) => ("", name),
+                    (Some(ns), name) => (ns, name),
+                };
+                update.add_event(UpdateEvent::AddNodeLabel {
                     node_name: node_name.to_string(),
-                    node_type: "corpus".to_string(),
-                })?; // this is required, corpus annotations might be first updates to be processed
-                if let Some(previous_name) = previous {
-                    update.add_event(UpdateEvent::AddEdge {
-                        source_node: node_name.to_string(),
-                        target_node: previous_name.to_string(),
-                        layer: ANNIS_NS.to_string(),
-                        component_type: AnnotationComponentType::PartOf.to_string(),
-                        component_name: "".to_string(),
-                    })?;
-                }
-                previous = Some(node_name);
-            }
-            if let Some(corpus_doc_path) = previous {
-                let path = file_path.as_path();
-                let annotations = read_annotations(path, &progress)?;
-                for (k, v) in annotations {
-                    let (anno_ns, anno_name) = match split_qname(k.as_str()) {
-                        (None, name) => ("", name),
-                        (Some(ns), name) => (ns, name),
-                    };
-                    update.add_event(UpdateEvent::AddNodeLabel {
-                        node_name: corpus_doc_path.to_string(),
-                        anno_ns: anno_ns.to_string(),
-                        anno_name: anno_name.to_string(),
-                        anno_value: v,
-                    })?;
-                }
+                    anno_ns: anno_ns.to_string(),
+                    anno_name: anno_name.to_string(),
+                    anno_value: v,
+                })?;
             }
             progress.worked(1)?;
         }
@@ -151,13 +120,31 @@ mod tests {
     }
 
     fn test_metadata(on_disk: bool) -> Result<(), Box<dyn std::error::Error>> {
-        let mut e_g = target_graph(on_disk)?;
+        let mut e_g = target_graph(on_disk).map_err(|_| assert!(false)).unwrap();
         let add_metadata = AnnotateCorpus::default();
-        let metadata = ["language=unknown", "date=yesterday"];
+        // document-level metadata
+        let doc_metadata = ["language=unknown", "date=yesterday"];
         let metadata_file_path = temp_dir().join("metadata").join("corpus").join("doc.meta");
-        std::fs::create_dir_all(metadata_file_path.parent().unwrap())?;
-        let mut metadata_file = std::fs::File::create(metadata_file_path)?;
-        metadata_file.write(metadata.join("\n").as_bytes())?;
+        std::fs::create_dir_all(metadata_file_path.parent().unwrap())
+            .map_err(|_| assert!(false))
+            .unwrap();
+        let mut metadata_file = std::fs::File::create(metadata_file_path)
+            .map_err(|_| assert!(false))
+            .unwrap();
+        metadata_file
+            .write(doc_metadata.join("\n").as_bytes())
+            .map_err(|_| assert!(false))
+            .unwrap();
+        // corpus-level metadata
+        let corpus_metadata = ["version=1.0", "doi=is a secret"];
+        let cmetadata_file_path = temp_dir().join("metadata").join("corpus.meta");
+        let mut cmetadata_file = std::fs::File::create(cmetadata_file_path)
+            .map_err(|_| assert!(false))
+            .unwrap();
+        cmetadata_file
+            .write(corpus_metadata.join("\n").as_bytes())
+            .map_err(|_| assert!(false))
+            .unwrap();
         let r = add_metadata.import_corpus(
             temp_dir().join("metadata").as_path(),
             add_metadata.step_id(None),
@@ -170,23 +157,37 @@ mod tests {
             r.err().unwrap()
         );
         let mut u = r?;
-        external_updates(&mut u)?;
-        let mut g = AnnotationGraph::new(on_disk)?;
+        external_updates(&mut u)
+            .map_err(|_| assert!(false))
+            .unwrap();
+        let mut g = AnnotationGraph::new(on_disk)
+            .map_err(|_| assert!(false))
+            .unwrap();
         let apu = g.apply_update(&mut u, |_| {});
-        assert_eq!(
-            true,
+        assert!(
             apu.is_ok(),
             "Applying updates ends with error: {:?}",
-            &apu
+            &apu.err()
         );
-        let queries = ["tok @* language", "tok @* date"];
+        let queries = [
+            "language",
+            "date",
+            "version",
+            "doi",
+            "annis:node_name=/corpus/ _ident_ version=/1.0/ _ident_ doi=/is a secret/",
+            "annis:node_name=\"corpus/doc\" _ident_ language=/unknown/ _ident_ date=/yesterday/",
+        ];
         let corpus_name = "current";
-        let tmp_dir_e = tempdir_in(temp_dir())?;
-        let tmp_dir_g = tempdir_in(temp_dir())?;
-        e_g.save_to(&tmp_dir_e.path().join(corpus_name))?;
-        g.save_to(&tmp_dir_g.path().join(corpus_name))?;
-        let cs_e = CorpusStorage::with_auto_cache_size(&tmp_dir_e.path(), true)?;
-        let cs_g = CorpusStorage::with_auto_cache_size(&tmp_dir_g.path(), true)?;
+        let tmp_dir_e = tempdir_in(temp_dir()).map_err(|_| assert!(false)).unwrap();
+        let tmp_dir_g = tempdir_in(temp_dir()).map_err(|_| assert!(false)).unwrap();
+        assert!(e_g.save_to(&tmp_dir_e.path().join(corpus_name)).is_ok());
+        assert!(g.save_to(&tmp_dir_g.path().join(corpus_name)).is_ok());
+        let cs_e = CorpusStorage::with_auto_cache_size(&tmp_dir_e.path(), true)
+            .map_err(|_| assert!(false))
+            .unwrap();
+        let cs_g = CorpusStorage::with_auto_cache_size(&tmp_dir_g.path(), true)
+            .map_err(|_| assert!(false))
+            .unwrap();
         for query_s in queries {
             let query = SearchQuery {
                 corpus_names: &[corpus_name],
@@ -194,13 +195,21 @@ mod tests {
                 query_language: QueryLanguage::AQL,
                 timeout: None,
             };
-            let matches_e = cs_e.find(query.clone(), 0, None, ResultOrder::Normal)?;
-            let matches_g = cs_g.find(query, 0, None, ResultOrder::Normal)?;
+            let matches_e = cs_e
+                .find(query.clone(), 0, None, ResultOrder::Normal)
+                .map_err(|_| assert!(false))
+                .unwrap();
+            let matches_g = cs_g
+                .find(query, 0, None, ResultOrder::Normal)
+                .map_err(|_| assert!(false))
+                .unwrap();
+            assert!(matches_e.len() > 0, "No matches for query: {}", query_s);
             assert_eq!(
                 matches_e.len(),
                 matches_g.len(),
-                "Failed with query: {}",
-                query_s
+                "Failed with query: {} ({:?})",
+                query_s,
+                matches_g
             );
             for (m_e, m_g) in matches_e.into_iter().zip(matches_g.into_iter()) {
                 assert_eq!(m_e, m_g);
@@ -211,7 +220,7 @@ mod tests {
 
     fn external_updates(u: &mut GraphUpdate) -> Result<(), Box<dyn std::error::Error>> {
         u.add_event(UpdateEvent::AddNode {
-            node_name: "corpus/doc#1".to_string(),
+            node_name: "corpus/doc#t1".to_string(),
             node_type: "node".to_string(),
         })?;
         u.add_event(UpdateEvent::AddNodeLabel {
@@ -286,12 +295,24 @@ mod tests {
             node_name: "corpus".to_string(),
             node_type: "corpus".to_string(),
         })?;
+        u.add_event(UpdateEvent::AddNodeLabel {
+            node_name: "corpus".to_string(),
+            anno_ns: "".to_string(),
+            anno_name: "version".to_string(),
+            anno_value: "1.0".to_string(),
+        })?;
+        u.add_event(UpdateEvent::AddNodeLabel {
+            node_name: "corpus".to_string(),
+            anno_ns: "".to_string(),
+            anno_name: "doi".to_string(),
+            anno_value: "is a secret".to_string(),
+        })?;
         u.add_event(UpdateEvent::AddNode {
             node_name: "corpus/doc".to_string(),
             node_type: "corpus".to_string(),
         })?;
         u.add_event(UpdateEvent::AddNode {
-            node_name: "corpus/doc#1".to_string(),
+            node_name: "corpus/doc#t1".to_string(),
             node_type: "node".to_string(),
         })?;
         u.add_event(UpdateEvent::AddNodeLabel {