+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed ordering of token nodes in spreadsheet import
 - `link` now considers all matching nodes for the same value, so the correct amount of edges is created
 - `exmaralda` returns error when there is no time value for a timeline item
+- fixed and simplified import of corpus node annotations
 
 ### Removed